[Performance] - Transform "geoserver_post_save_layers" to an asynchronous task
